+++ src/map.c
@@ -1,9 +1,8 @@
 #include <stdlib.h>
-#include <inttypes.h>
 #include "map.h"
 #include "tile.h"
 
-void map_init( uint16_t width, uint16_t height, uint16_t depth )
+void map_init( int width, int height, int depth )
 {
   	map.width = width;
 	map.height = height;
